Address peer feedback
