+++ docs/csharp/programming-guide/concepts/linq/snippets/extensions/LinqExtensions.cs
@@ -1,24 +1,21 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace linq
+namespace Custom.Linq.Extensions
 {
     // <LinqExtensionClass>
-    public static class LINQExtension
+    public static class EnumerableExtension
     {
         public static double Median(this IEnumerable<double>? source)
         {
-            if (!(source?.Any() ?? false))
+            if (source is null || !source.Any())
             {
                 throw new InvalidOperationException("Cannot compute median for a null or empty set.");
             }
 
-            var sortedList = (from number in source
-                              orderby number
-                              select number).ToList();
+            var sortedList =
+                source.OrderBy(number => number).ToList();
 
             int itemIndex = sortedList.Count / 2;
 
@@ -39,15 +36,15 @@ namespace linq
     public static class OtherExtensions
     {
         // <IntOverload>
-        //int overload
+        // int overload
         public static double Median(this IEnumerable<int> source) =>
-            (from num in source select (double)num).Median();
+            (from number in source select (double)number).Median();
         // </IntOverload>
 
         // <GenericOverload>
-        // Generic overload.
-        public static double Median<T>(this IEnumerable<T> numbers,
-                               Func<T, double> selector) =>
+        // generic overload
+        public static double Median<T>(
+            this IEnumerable<T> numbers, Func<T, double> selector) =>
             (from num in numbers select selector(num)).Median();
         // </GenericOverload>
 
@@ -56,18 +53,17 @@ namespace linq
         // The method returns every other element of a sequence.
         public static IEnumerable<T> AlternateElements<T>(this IEnumerable<T> source)
         {
-            int i = 0;
-            foreach (var element in source)
+            int index = 0;
+            foreach (T element in source)
             {
-                if (i % 2 == 0)
+                if (index % 2 == 0)
                 {
                     yield return element;
                 }
-                i++;
+
+                index++;
             }
         }
         // </SequenceElement>
-
-
     }
 }
