+++ docs/csharp/programming-guide/concepts/linq/snippets/extensions/Program.cs
@@ -1,0 +1,63 @@
+﻿using Custom.Linq.Extensions;
+
+// <MedianUsage>
+double[] numbers = { 1.9, 2, 8, 4, 5.7, 6, 7.2, 0 };
+var query = numbers.Median();
+
+Console.WriteLine($"double: Median = {query}");
+// This code produces the following output:
+//     double: Median = 4.85
+// </MedianUsage>
+
+// <OverloadUsage>
+double[] numbers1 = { 1.9, 2, 8, 4, 5.7, 6, 7.2, 0 };
+var query1 = numbers1.Median();
+
+Console.WriteLine($"double: Median = {query1}");
+
+int[] numbers2 = { 1, 2, 3, 4, 5 };
+var query2 = numbers2.Median();
+
+Console.WriteLine($"int: Median = {query2}");
+// This code produces the following output:
+//     double: Median = 4.85
+//     int: Median = 3
+// </OverloadUsage>
+
+// <GenericUsage>
+int[] numbers3 = { 1, 2, 3, 4, 5 };
+
+/*
+    You can use the num => num lambda expression as a parameter for the Median method
+    so that the compiler will implicitly convert its value to double.
+    If there is no implicit conversion, the compiler will display an error message.
+*/
+var query3 = numbers3.Median(num => num);
+
+Console.WriteLine($"int: Median = {query3}");
+
+string[] numbers4 = { "one", "two", "three", "four", "five" };
+
+// With the generic overload, you can also use numeric properties of objects.
+var query4 = numbers4.Median(str => str.Length);
+
+Console.WriteLine($"string: Median = {query4}");
+// This code produces the following output:
+//     int: Median = 3
+//     string: Median = 4
+// </GenericUsage>
+
+// <SequenceUsage>
+string[] strings = { "a", "b", "c", "d", "e" };
+
+var query5 = strings.AlternateElements();
+
+foreach (var element in query5)
+{
+    Console.WriteLine(element);
+}
+// This code produces the following output:
+//     a
+//     c
+//     e
+// </SequenceUsage>
